Update README.md
In this sentence: "The key of this project is to help you understand how to pick the components that you want to build, you have to think the best way to organize and avoid future errors. What can be re-used? If your code is tuning into something big and hard to understand you probably will need to split it into components:" The word 'tuning' should be 'turning'. It's a spelling mistake; it's missing an 'r'.